feat: implement firebase and add entries to cloud storage

+++ lib/models/month_data.dart
@@ -2,7 +2,7 @@ import 'package:flutter_finances/utils/masks.dart';
 
 class MonthData {
   String type;
-  int value;
+  double value;
   String date;
 
   MonthData({required this.type, required this.value, required this.date});
